Introduce nested schemas and guarded rules
This adds a new interface that will create nested schemas within a
schema, ie:

  Dry::Validation.Schema do
    key(:location).schema do
      key(:lat).required(:float?)
      key(:lng).required(:float?)
    end
  end

This is the first step towards support for reusable schemas but it
actually comes with sweet benefits, with nested schemas generating error
messages is simpler, because we don't have to figure out the structure
from a huge result ast.

In addition to that we have now guarded rules, which you can define with
a new syntax:

  Dry::Validation.Schema do
    key(:bar).required
    key(:baz).required

    rule(foo: [:bar, baz]) do |bar, baz|
      bar.eql?('such feature') & baz.eql?('very wow')
    end
  end

This is a *huge* improvement as we can safely define rules that will not
be executed unless their dependencies are valid according to schema
specification. In this example if, let's say, `:baz` key is missing, our
`:foo` rule will not crash as it'll be not executed since its deps are
invalid.

refs #62