Make Schema quack like a rule wrt currying
This subtle addition makes schemas work with high-level rules which are
handled by `Rule::Check` from dry-logic. The requirement was to not only
respond to `call` and return a dry-logic's result-compatible object, but
also to support currying args, so that we can partially apply additional
args and then apply final input for validation, which is how every
rule predicate works.

This is a step forward towards making schemas work like rules, which
should make the whole system more flexible as schemas are composable now
just like rules.