Support interpolation high-lvl rule messages
Previously error visitor would stop message compilation when it found a
message for a custom rule (typically a high-lvl rule), so we wouldn't be
able to extract any potential tokens for interpolation.

This change makes the compiler go to down to the predicate even when
there's a custom message for a custom rule name, and process it like the
built-in messages, which means we get interpolation working too.