Update to work with new dry-logic and dry-struct
Apologies for this gigantic commit but it wasn't possible to update this
without making it work with dry-types too as we already moved its Struct
api to dry-struct, and I also had to update dry-types to use new
dry-logic APIs (still wip over there but bare minimum was done to make
dry-v specs passing).

Anyhow, this commit is *mostly removing code* as dry-logic has been
internally refactored in a way that made various things much simpler in
dry-v, furthermore some messages that we were not able to generate
before are now fully working, so various pending specs are now passing.

The public API *does not change*, however, some messages are now
different, but it's OK since previously they were not 100% accurate (see
previously pending specs that are now enabled).

More work has to be done in master now so that we can support generating
messages for all rules and operations, specifically `or` and `xor` operations
are not fully supported yet, which will be added in master.

Awesome bonus of this refactor: ~2x performance boost and drastically
simplified message compilation.

Summary:

  * Remove `HintCompiler` as hints are now part of result AST
  * Remove `ErrorCompiler` and move its methods to `MessageCompiler`
  * Update `MessageCompiler` to work with the new result AST
  * Remove custom `:error` nodes from result AST as they are no longer
    needed
  * Remove `Error` wrappers
  * Remove `BuildError` step from executor as we no longer need to wrap
    results with custom objects
  * Remove `Dry::Logic::Predicate#eval_args!` patch in favor of the new
    `Rule::Predicate#eval_args` method
  * Remove generation of `:val` node as they are no longer needed
  * Update specs that were using `predicate { .. }` syntax for defining
    custom predicates and simply use method definitions
  * Update to work with `Dry::Struct`
  * Enable pending specs as they are passing now \o/
  * Mark one spec as pending as its expectation was broken anyway - full
    support for "or" messages will be added shortly in master