Refine behavior of messages for high-level rules and custom rules
This is a breaking change as it removes support for when macro with
multiple rules inside, which means that if you use `when` then its block
must return a rule. This is fine though as `when` should only be used
for simple checks anyway and I'm pretty sure barely anyone was using it
to define multiple rules. Using `rule` or `validate` is better anyway
and honestly I'd vote for removing `when` macro in 1.0.0. Something to
discuss...

Anyhow, this change restores behavior from 0.9.x where explicitly named
high-level rule will result in messages attached to that name, which is
good since we're not breaking that behavior after all. Furthermore, the
exact same behavior is present with the new `validate` feature, which
means that its name is used for messages.

The name *will not be used* for messages if it happens to be a predicate
name (read: ends with "?"), in such case we will use the name from the
value path, like for any other rule.

Closes #250