builds better schema rules from Dry::Struct
Dry::Struct can be used in schema validation with the `.schema` method
to define a nested structure that should validate (satisfy) a predefined
Dry::Struct.

The issue addressed here is that there are some assumptions baked into
the StructClassBuilder that has several unintuitive effects:

1. all keys are declared as `required`, which ignores the
meta[:omittable] metadata declared on the struct field type
2. all keys' values are assumed as `.filled`, which may not be the case
and ignores the Struct field's Type coercible and `optional` settings
3. Structs with Array-type fields don't validate the array members properly, as
noted in #396 (https://github.com/dry-rb/dry-validation/pull/396)
4. Structs with Array-type fields aren't allowed to be empty, which a
Struct allows.

I'm in agreement about not using structs to validate, but it seems like
this functionality was at least somewhat intended as the extension
exists today. If this is unwanted there's already a workaround--define
your schema validation separately from your struct. Doing so means
a good deal of duplicating the type information between the schema and
the struct.

This modifies the StructClassBuilder to build a schema more
intelligently, by respecting `omittable` and handling arrays of types
and other nested structs. The time complexity seems similar in that this
is as recursive as the previous implementation, with the exception that
this has a few easy conditionals.