Remove paths from Rule
I realized it's going to be much better to expose a nice API for reading
values on dry-schema side, spefifically `Result` object can provide
various convenience methods for quickly accessing values.

This way we can simplify rule API here, as it just needs an identifier,
and params are accessible within the blocks already, so we're all set.

Another benefit is that `failure` method doesn't need the rule
identifier as it can be provided behind the scenes from the rule object
itself.