Make Result#error?(key) return true for rule and schema errors
Not super happy with this as of now :) Exchanged usage of
error? with schema_error? in the code base as otherwise there
thankfully were a bunch of failures.

The implementation of error? itself is basically stolen from
its counterpart at `Schema::Result`. It makes sense to check
if there are any errors for this but not sure if we shouldn't
rather still call out to schema_result.error?

Also, root/base behavior is a bit of a mystery.