Evaluator#key? forwards key name argument
When testing the presence of multiple, optional keys, it would be nice
to be able to specify the key name to `key?` as you can with
`values.key?`.

Since this merely delegates to that method, I can't see any reason why
this omission was intentional. `#key_name` is preserved as the default
argument, so this will not change any existing behavior.